+++ src/api/api.ts
@@ -1,33 +1,22 @@
-import { useQuery } from 'react-query';
-import { axiosInstance } from './clients';
-import type { Chapter, Definition, ID } from './types';
+import axios from 'axios';
+
+export const baseURL = import.meta.env.BASE_URL || '/';
+
+const axiosInstance = axios.create({
+  baseURL,
+  headers: {
+    'Content-Type': 'application/json',
+    Accept: 'application/json',
+  },
+});
 
 const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
 
-async function fetchData(url: string) {
+export async function fetchData(url: string) {
   const data = await axiosInstance.get(url).then((res) => res.data);
 
   // TODO: remove sleep
   // await sleep(3000);
 
   return data;
-}
-
-export function useChapters() {
-  return useQuery<Chapter[], Error>('chapters', () =>
-    fetchData('/chapters.json'),
-  );
-}
-
-export function useDefinitions(chapterId?: ID | undefined) {
-  if (chapterId === undefined)
-    return useQuery<Definition[], Error>('definitions', () =>
-      fetchData('/definitions.json'),
-    );
-
-  return useQuery<Definition[], Error>(['definitions', chapterId], () =>
-    fetchData('/definitions.json').then((definitions: Definition[]) =>
-      definitions?.filter((definition) => definition.chapterId === chapterId),
-    ),
-  );
 }
