--- src/api/clients.ts
@@ -1,21 +1,0 @@
-import { QueryClient } from 'react-query';
-import axios from 'axios';
-
-export const baseURL = import.meta.env.BASE_URL || '/';
-
-export const axiosInstance = axios.create({
-  baseURL,
-  headers: {
-    'Content-Type': 'application/json',
-    Accept: 'application/json',
-  },
-});
-
-export const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      refetchOnWindowFocus: false,
-      refetchOnMount: false,
-    },
-  },
-});
