+++ src/api/index.ts
@@ -1,3 +1,6 @@
 export * from './api';
-export * from './clients';
+export * from './chapters';
+export { default as queryClient } from './client';
+export * from './definitions';
+export * from './favourites';
 export * from './types';
