+++ src/components/card/DefinitionCard.tsx
