+++ src/components/card/index.ts
@@ -1,0 +1,2 @@
+export { default as Collection } from './Collection';
+export { default } from './DefinitionCard';
