+++ src/components/chapter/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Chapter';
+export { default as ChapterHeader } from './ChapterHeader';
