+++ src/components/feedback/index.ts
@@ -1,0 +1,1 @@
+export { default as FeedbackForm } from './FeedbackForm';
