+++ src/components/navbar/NavbarHeader.tsx
@@ -1,0 +1,11 @@
+import Typography from '@mui/material/Typography';
+
+function NavbarHeader() {
+  return (
+    <Typography variant="subtitle1" component="h2" sx={{ flexGrow: 1 }}>
+      Made with ❤️ by Haltarys!
+    </Typography>
+  );
+}
+
+export default NavbarHeader;
