+++ src/components/navbar/SourceCodeButton.tsx
@@ -1,0 +1,20 @@
+import Tooltip from '@mui/material/Tooltip';
+import IconButton from '@mui/material/IconButton';
+import CodeIcon from '@mui/icons-material/Code';
+
+function SourceCodeButton() {
+  return (
+    <Tooltip title="View the source code on Github!">
+      <IconButton
+        size="large"
+        color="inherit"
+        href="https://github.com/Haltarys/HIST130-Quiz"
+        target="_blank"
+      >
+        <CodeIcon />
+      </IconButton>
+    </Tooltip>
+  );
+}
+
+export default SourceCodeButton;
