+++ src/components/navbar/index.ts
@@ -1,0 +1,5 @@
+export { default as NavbarHeader } from './NavbarHeader';
+export { default as NavbarTabs } from './NavbarTabs';
+export { default as NavbarTitle } from './NavbarTitle';
+export { default as SourceCodeButton } from './SourceCodeButton';
+export { default } from './Navbar';
