+++ src/layout/index.ts
@@ -1,0 +1,2 @@
+export { default as Navbar } from 'src/components/navbar';
+export { default as Body } from './Body';
