+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export { default as Favourites } from './Favourites';
+export { default as Quiz } from './Quiz';
+export { default as Reference } from './Reference';
