+++ src/store/index.ts
@@ -1,3 +1,2 @@
 export * from './Store';
-import StoreProvider from './Store';
-export default StoreProvider;
+export { default } from './Store';
